Use separate for public headers
