+++ mir-tests/loop-sieve-gen.c
@@ -1,4 +1,4 @@
-#include "../mir-gen.h"
+#include <mir-gen.h>
 #include "../real-time.h"
 #if defined(TEST_GEN_LOOP)
 #include "api-loop.h"
